Unify util folder (#244)
