+++ pipeline/cutting.py
@@ -1,7 +1,7 @@
 
 from pipeline.parameters import Parameters
 from pipeline.data_store import DataStore
-from pipeline.utils import parse_session
+from util.util import parse_session
 
 from models.source_list import SourceList
 from models.cutouts import Cutouts
